+++ lib/changelog/parsers.rb
@@ -1,5 +1,6 @@
 require 'changelog/parsers/base'
 require 'changelog/parsers/intridea_format'
+require 'changelog/parsers/keep_a_changelog'
 
 module Danger
   module Changelog
@@ -9,7 +10,7 @@ module Danger
       end
 
       def self.lookup(format)
-        { intridea: IntrideaFormat }
+        { intridea: IntrideaFormat, keep_a_changelog: KeepAChangelog }
           .fetch(format, IntrideaFormat)
           .new
       end